fix: api methods 이동
