+++ src/App.js
@@ -36,11 +36,6 @@ function App() {
       </Router>
     </div>
   );
-  // return (
-  //   <div>
-  //     <PostListFilter />
-  //   </div>
-  // );
 }
 
 export default App;
