+++ src/api/auth-api.js
@@ -1,13 +1,5 @@
 import axios from "axios";
-import {
-  AUTH_USER,
-  BASE_URL,
-  LOGIN,
-  LOGOUT,
-  SIGNUP,
-  UPLOAD_PROFILE,
-  UPLOAD_COVER,
-} from "./url";
+import { AUTH_USER, BASE_URL, LOGIN, LOGOUT, SIGNUP } from "./url";
 
 export const userLogin = async (email, password) => {
   const res = await axios.post(`${BASE_URL}${LOGIN}`, {
@@ -34,27 +26,6 @@ export const userAuth = async (token) => {
   const res = await axios.get(`${BASE_URL}${AUTH_USER}`, {
     headers: {
       Authorization: `${token}`,
-    },
-  });
-
-  return res;
-};
-
-export const uploadCoverImage = async (file, token) => {
-  const res = await axios.post(`${BASE_URL}${UPLOAD_COVER}`, file, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-      "content-type": "multipart/form-data",
-    },
-  });
-  return res;
-};
-
-export const uploadProfileImage = async (file, token) => {
-  const res = await axios.post(`${BASE_URL}${UPLOAD_PROFILE}`, file, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-      "Content-Type": "multipart/form-data",
     },
   });
